Move bank_iterator flow to __main__
